test(android-e2e): reveal automation canaries by visibility probe
`smoke:automation-system` revealed its canaries with a single blind
`scroll down <amount>` and then acted on them immediately. `scroll` is a
gesture, not an offset — the command's own contract says "app scroll
physics determine the final content offset" — so one amount cannot
guarantee the target is on screen, least of all right after the
landscape/portrait round-trip has relaid the list out.

Both open Android smoke failures sit on that seam:

- step 34 `press id="automation-press"` → `selector_not_found`, right
  after `scroll down 0.7`. Reproduced on #2360 and #2361 (disjoint
  diffs), on `main` at bd08e6e, and on #2356 at the identical line, so
  the orientation fix does not cover it.
- the post-alert canary after `scroll down 1`, which #2356 records as
  still open.

The iOS scenario already reveals the same canaries correctly, probing
`is visible` and scrolling again until the element is on screen. That
policy was iOS-only; this moves it to the shared live-device layer as
`createVisibilityScroll` and gives the Android scenario the same reveal.
Its unit test moves with it. No production code changes, and the press
and diff assertions still have to pass against a real element — the
scroll budget is bounded and a genuinely absent canary still fails.
